fix: RadarChart tick and graph point scale do not match
Fixes an issue where RadarChart tick and graph point scale do not match.

If a point has a value of 15 and a tick line has a value of 15, I would expect the point to fall on the line. This was not always the case.

The values in tick titles did not always match with the values of the points on the graph because
- points were drawn assuming the graph center = 0
- tick title values assumed that the graph center = (minValue - ((maxValue - minValue) / tickCount))

This commit resolves this issue by changing the scaling factor applied to points when they are drawn so that they match the scale of the tick lines.

New unit tests have been added and existing tests that used the incorrect scaling have been updated.

Resolves #1078